src/grpc/mpes-stt/requirements.txt
fastapi
uvicorn
openai_whisper
python-multipart
grpcio
grpcio-tools
protobuf